# method_pipeline/generated/chatdev/minimax-m2.7/generic/QA2-1207/sdk_chatdev_20260712091919_20260712091919/code_workspace/oms_backend/openapi.yaml
openapi: 3.0.3
info:
  title: Order Management System API
  description: |
    A production-grade Order Management System backend implementing the complete order workflow.
    
    ## Workflow Steps
    1. Customer places order
    2. Order Staff reviews & accepts
    3. Accountant creates invoice
    4. Customer pays invoice
    5. Accountant verifies payment
    6. Order Staff ships paid order
    7. Order Staff closes completed order
    
    ## NFR Support
    - NFR 2.1 Graceful Degradation: Feature flags and circuit breakers
    - NFR 2.2 Fault Detection: Health checks and automatic recovery
    - NFR 2.3 State Preservation: WAL mode, snapshots, and idempotency keys
  version: 1.0.0
  contact:
    name: OMS Backend Team
servers:
  - url: http://localhost:8000
    description: Local development server
paths:
  /api/v1/customers:
    get:
      tags: [customers]
      summary: List customers
      parameters:
        - name: skip
          in: query
          schema: {type: integer, default: 0}
        - name: limit
          in: query
          schema: {type: integer, default: 100}
      responses:
        '200':
          description: List of customers
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Customer'
    post:
      tags: [customers]
      summary: Create customer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateCustomerRequest'
      responses:
        '201':
          description: Customer created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Customer'
  /api/v1/customers/{customer_id}:
    get:
      tags: [customers]
      summary: Get customer by ID
      parameters:
        - name: customer_id
          in: path
          required: true
          schema: {type: string}
      responses:
        '200':
          description: Customer found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Customer'
        '404':
          description: Customer not found
    patch:
      tags: [customers]
      summary: Update customer
      parameters:
        - name: customer_id
          in: path
          required: true
          schema: {type: string}
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateCustomerRequest'
      responses:
        '200':
          description: Customer updated
    delete:
      tags: [customers]
      summary: Delete customer
      parameters:
        - name: customer_id
          in: path
          required: true
          schema: {type: string}
      responses:
        '204':
          description: Customer deleted
  /api/v1/products:
    get:
      tags: [products]
      summary: List products
      parameters:
        - name: skip
          in: query
          schema: {type: integer, default: 0}
        - name: limit
          in: query
          schema: {type: integer, default: 100}
        - name: active_only
          in: query
          schema: {type: boolean, default: false}
      responses:
        '200':
          description: List of products
    post:
      tags: [products]
      summary: Create product
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateProductRequest'
      responses:
        '201':
          description: Product created
  /api/v1/products/{product_id}:
    get:
      tags: [products]
      summary: Get product by ID
      parameters:
        - name: product_id
          in: path
          required: true
          schema: {type: string}
      responses:
        '200':
          description: Product found
...
  /api/v1/invoices:
    post:
      tags: [invoices]
      summary: Create invoice (Step 3)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateInvoiceRequest'
      responses:
        '201':
          description: Invoice created
  /api/v1/invoices/{invoice_id}:
    get:
      tags: [invoices]
      summary: Get invoice by ID
      parameters:
        - name: invoice_id
          in: path
          required: true
          schema: {type: string}
      responses:
        '200':
          description: Invoice found
...
components:
  schemas:
    Customer:
      type: object
      properties:
        id: {type: string}
        name: {type: string}
        email: {type: string}
        phone: {type: string}
        address: {$ref: '#/components/schemas/Address'}
        role: {type: string}
        created_at: {type: string}
        updated_at: {type: string}
    Address:
      type: object
      properties:
        street: {type: string}
        city: {type: string}
        state: {type: string}
        postal_code: {type: string}
        country: {type: string}
    CreateCustomerRequest:
      type: object
      required: [name, email]
      properties:
        name: {type: string}
        email: {type: string}
        phone: {type: string}
        address: {$ref: '#/components/schemas/Address'}
    UpdateCustomerRequest:
      type: object
      properties:
        name: {type: string}
        email: {type: string}
        phone: {type: string}
...
    CreateProductRequest:
      type: object
      required: [sku, base_price]
      properties:
        sku: {type: string}
        description: {type: string}
        base_price: {type: number}
        currency: {type: string, default: USD}
        stock_quantity: {type: integer, default: 0}
...
    CreateInvoiceRequest:
      type: object
      required: [order_id, customer_id, billing_address]
      properties:
        order_id: {type: string}
        customer_id: {type: string}
        billing_address: {$ref: '#/components/schemas/Address'}
        due_date_days: {type: integer, default: 30}
...